refactor(link): use flat structure
