+++ src/__stories__/index.ts
@@ -11,9 +11,9 @@ import './Icon';
 import './Image';
 import './Input';
 import './KitchenSink';
+import './Link';
 import './List';
 import './Menu';
-import './SlateEditor/Link';
 import './Table';
 import './Text';
 import './Textarea';
